Implemented Eulerian results extractor based on path method
Path method creates a very short path that passes the location of interest. The path is used to extract field ouptut value using built-in plot by path method. This data extraction process is repeated for every frame so a time history of field output at the spatial location is obtained. This method bypasses the cumbersome process to find the closest nodes or elements to the spatial location, thus, has a much higher process efficiency.