fixed writing odb permission
Previously, odb files are open in writable mode, some node and element sets created in the middle of execution are writen into odb files which may contradict when the script runs on the same odb file with different inputs for a later time.
Change open odb to read only mode.